+++ src/main/java/bit/user/oauth/kakao/service/KaKaoLoginServiceImpl.java
@@ -1,48 +1,24 @@
-package bit.user.oauth.kakao.controller;
+package bit.user.oauth.kakao.service;
 
+import bit.user.oauth.port.OAuthService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/oauth")
-public class KaKaoController {
+@Service
+public class KaKaoLoginServiceImpl implements OAuthService {
 
-    @Value("${kakao.client.id}")
-    String clientId;
-    @Value("${kakao.redirect.uri}")
-    String redirectUri;
-    @Value("${kakao.client.secret}")
-    String clientSecret;
+    @Override
+    public String getToken(String code, String clientId, String redirectUri, String clientSecret)
+            throws JsonProcessingException {
 
-    @GetMapping("/kakao")
-    public String getAuthKakao() {
-        return "https://kauth.kakao.com/oauth/authorize?"
-                + "client_id=" + clientId
-                + "&redirect_uri=" + redirectUri
-                + "&response_type=code"
-                + "&scope=account_email";
-    }
-
-    @PostMapping("/kakao/token")
-    public String postAuthKakao(@RequestBody MultiValueMap<String, String> params) throws JsonProcessingException {
-        String code = params.getFirst("code");
         HttpHeaders headers = new HttpHeaders();
         headers.add("Content-Type", "application/x-www-form-urlencoded;charset=UTF-8");
         MultiValueMap<String, String> httpBody = new LinkedMultiValueMap<>();
@@ -64,8 +40,8 @@ public class KaKaoController {
         return jsonNode.get("access_token").asText();
     }
 
-    @PostMapping("/kakao/access")
-    public String postGetUserInfo(@RequestBody String accessToken) throws JsonProcessingException {
+    @Override
+    public String getUserInfo(String accessToken) {
         HttpHeaders headers = new HttpHeaders();
         headers.add("Authorization", "Bearer " + accessToken);
         headers.add("Content-Type", "application/x-www-form-urlencoded;charset=UTF-8");
@@ -75,7 +51,6 @@ public class KaKaoController {
         ResponseEntity<String> response = restTemplate.postForEntity(
                 "https://kapi.kakao.com/v2/user/me", httpEntity, String.class
         );
-
         return response.getBody();
     }
 }
